+++ src/GameSrv/Monster/Monsters/SpitSpider.cs
@@ -25,9 +25,9 @@ namespace GameSrv.Monster.Monsters {
                         short nY = (short)(CurrY - 2 + i);
                         BaseObject baseObject = Envir.GetMovingObject(nX, nY, true);
                         if (baseObject != null && baseObject != this && IsProperTarget(baseObject) && M2Share.RandomNumber.Random(baseObject.SpeedPoint) < HitPoint) {
-                            nDamage = baseObject.GetMagStruckDamage(this, (ushort)nDamage);
+                            nDamage = baseObject.GetMagStruckDamage(this, nDamage);
                             if (nDamage > 0) {
-                                baseObject.StruckDamage((ushort)nDamage);
+                                baseObject.StruckDamage(nDamage);
                                 baseObject.SendStruckDelayMsg(Messages.RM_STRUCK, Messages.RM_REFMESSAGE, nDamage, WAbil.HP, WAbil.MaxHP, ActorId, "", 300);
                                 if (UsePoison) {
                                     if (M2Share.RandomNumber.Random(AntiPoison + 20) == 0) {
